requirements.txt
fastapi
uvicorn
sqlmodel
python-jose
passlib[bcrypt]
psycopg2-binary
pydantic-settings
python-multipart
requests
google-auth
email-validator
